准备将 深度缓存渲染到纹理，开始准备渲染策略

+++ DZSmaEngine/Engine/Runtime/Engine/Private/RenderProxy/SimpleRenderProxy.cpp
@@ -18,11 +18,11 @@ void SimpleMeshRenderProxy::InitRender()
 	
 	mVertexBuffer = RHICreateVertexBuffer(mVertexData.data(),
 		sizeof(VertexPosNormalTangentTex),
-		mVertexData.size()
+		(UINT)mVertexData.size()
 	);
 	mIndexBuffer = RHICreateIndexBuffer(
 		mIndexData.data(),
-		mIndexData.size()
+		(UINT)mIndexData.size()
 	);
